implement implicit items/pairs iterators (#962)
Behaves as in [original
Nim](https://github.com/nim-lang/Nim/blob/d4098e6ca031aba9825980f9a17d3b54a9990577/compiler/semstmts.nim#L1297-L1305):
whether to use `items` or `pairs` is decided by the number of variables
not wrapped in `()`, and the error for the overloads is kept. I expected
the original compiler to restrict the overloads to `iterator` symbols
but it doesn't.

The iterator call check (which has to be performed again here) now makes
sure the produced expression is a call. The problem with checking for
identifiers for concepts is it could be a proc call from a concept and
not an iterator. Maybe we could fix this and make the check easier by
producing a different tag like `IcallX` for iterator calls or wrapping
the call in it, this might help with things like typechecking `toSeq`
too.