split `enum` typeclass into ordinal and holey type kinds (#973)
refs #968, closes #952

`enum` is split into 2 typeclasses: `OrdinalEnum` and `HoleyEnum`, which
compile to `(typekind (enum))` and `(typekind (holeyenum))`. For
simplicity, what previously parsed as the `enum` typeclass now parses as
an identifier `enum` which is defined in system as `OrdinalEnum |
HoleyEnum`.

However since `or` types cannot match themselves yet (#931), some
overloads for `enum` have to be temporarily split into overloads for
`OrdinalEnum` and `HoleyEnum`. The overloads for `==`, `<` etc though
are split so that generic disambiguation works since they are also
defined for `Ordinal`: `Ordinal` can be disambiguated with `OrdinalEnum`
since `OrdinalEnum` matches `Ordinal` but not vice versa, however it
cannot be disambiguated with `enum` in general since neither typeclass
matches one another. The other overloads should be able to be merged
when #931 or something similar is implemented, though I have not tested
this yet.