implement basic `when` inside object (#994)
fixes #965

The same code is reused for both regular `when` statements and `when`
statements inside objects, just with different endpoints for the
resolved expressions. So the way of handling unresolved conditions
depending on generic params etc. could be shared between the two
variants when implemented.

Although the real problem is if the `when` statement is left unresolved,
then expressions that iterate over the object type AST will encounter
`when` trees. The fields inside them would currently be untyped as well.
Nothing is done for this yet.