implement sum of products form for typeclass matching (#995)
follows up #931

Algorithm seems to work but I did not try every test case (cannot find
much on this,
[typescript](https://github.com/microsoft/TypeScript/blob/027829fbcd0a84172ab4a36d60a0ac915ca7ea09/src/compiler/checker.ts#L8467)
seems to do the distributive law but not de morgan since typescript
doesn't have `not` types). Since reordering can get quite slow
(reordering `(and (or <M terms>) (or <N terms>)` is `O(MN)`) the
typeclass is checked if it is already ordered first.

A problem with performance is the complexity of the produced terms also
reflects on the complexity of the type match, `(A or B) and (C or D) and
(E or F)` expands to:

```
(or
 (and A C E)
 (and A C F)
 (and A D E)
 (and A D F)
 (and B C E)
 (and B C F)
 (and B D E)
 (and B D F))
```

which performs 24 individual matches for only 6 types in the worst case.
`A and SomeInteger` for example matches `A` equally as many times as the
number of types in `SomeInteger`. An idea would be to cache the match
results but getting a key for the type and looking it up might be slower
than just matching.

This PR isn't really a priority, I just wanted to get the implementation
out of the way