implement case objects by just flattening fields (#996)
refs #992

Case objects are implemented by pretending their fields are flattened
and generating them as such in NIFC. The places that would need updating
to implement the normal behavior of case objects are (will include these
in #992):

1. destructor generation
2. NIFC generation (i.e. generating unions etc, also impacts object
constructor generation)
3. object constructors in sem, whatever they do in original nim
4. `sizeof` calculation

The AST representation is just a normal `case` statement where each of
the branches have a `(stmts)` body (nim parser does not generate
`nkRecList` without an indent so this gets inserted).

It would be great to get rid of `nil` or `discard` (not encountered I
think) in favor of just `.` but this is annoying to do in nifler, they
can be nested inside `nkRecList` etc.