basic generic/ref inheritance (#1040)
refs #896

Inheritance is implemented in sigmatch for `ref`/`ptr` and generic
versions of object types. This is not perfectly implemented for edge
cases like "generic objects matching generic objects" but should handle
the most common cases fine.

Changes to vtables_backend:

* Upcasts for ref/ptr types are just `cast`s of their pointers, this
works since the first field is always the parent object and makes the
refcount field for `ref`s work.
* Access to the vtable field needs an upcast to the hierarchy root
* Misc fixes: Temps are fixed (leaves `(expr` open instead of
immediately closing, generates `ptr` type for `addr` case), generic
types do not generate unused vtables, empty methods array generates `{
NIL }` since it is a flexible array.

Leaving for followup, added to issue todo and as XXX comments:

* Generic methods need to be instantiated for every single instantiation
of the attached generic type like hooks are at the end of sem, in the
test this is worked around by the methods matching statically for each
instance.
* Nil checks are needed for method dispatch and for `of` with pointer
type operands. Method dispatch is straightforward but I don't fully
understand what `of` does in the old compiler, the codegen implies it
should always return `false` when the operand is nil but it behaves like
a static match instead. In any case this would probably pollute the
diff.