instantiate methods for generic types (#1042)
For every generic type instantiation in a module, if any methods are
attached to the generic base type, the methods are instantiated for the
type instantiation as well (if they were not already). This is also
considered for local instantiations of generic types from other modules,
whose methods are now loaded from the index of the imported module.

Misc fixes: The index loading for methods was missing a ParRi skip.
Sigmatch for object type invocations forgot a skip. `{.untyped.}`
generic routines with return types are fixed, since they do not declare
`result` during the first semcheck they declare it during instantiation.