fixes for ref/ptr inheritance (#1049)
1. `ref`/`ptr` is skipped for the parent type of objects in hexer. Sem
could just generate these skipped but it might be better not to depend
on it.
2. In sigmatch, object conversions to generic types can be generated, to
instantiate these `genericConverter` is reused. (Non-pointer object
conversions ignored the type for upcasts so the type not being
instantiated wasn't a problem before)
3. Method instantiations for type instantiations skip modifiers and
`ref`/`ptr` when matching the method param, since the type instantiation
is always the object type.