add optional inheritance depth to `(kv)` in object constructors (#1057)
refs #896

Adds an optional inheritance level to `(kv)` in object constructors,
which NIFC uses in the designated initializer to add `.Q.Q...` to set a
field in the parent. A bit weird to use `(kv)` like this, but a new tag
is probably not worth it.

Originally created nested object constructors to take advantage of [NIFC
using them as the parent object
constructor](https://github.com/nim-lang/nimony/blob/bd72cd6d7393bbb1dabc7585983ef6a8aeb0f851/src/nifc/genexprs.nim#L318-L323),
but this resets the vtable pointer in the initializer.