fix forward declarations defined and used inside `when` bodies (#1062)
`when` statements are not processed until `SemcheckSignatures` due to
consts etc not being evaluated until then. However this meant that the
statements inside the `when` body only started with the phase
`SemcheckSignatures` and did not undergo the `SemcheckToplevelSyms`
phase, so forward declarations inside them were not processed. This is
fixed by doing every phase before `SemcheckSignatures` for the `when`
statement body if we are in `SemcheckSignatures`, and only doing the
current phase in the later phases.

I thought that not resolving `when` statements during
`SemcheckToplevelSyms` was a bigger problem than it was, proc forward
declarations still work, just types don't, which would not have been
possible in original Nim anyway.

```nim
type Foo = Bar

# works:
type Bar = int
# does not work:
when true:
  type Bar = int
```

Though in the future we could make this work for cases where the
compiler can understand that the condition is simple enough to compile
in `SemcheckToplevelSyms`.