implement {.inject.} in typed templates (#1077)
refs #1044 (the title, the full code in the issue needs untyped
parameters to work)

`{.inject.}` in untyped templates is implemented by just generating an
identifier instead of a sym, but this cannot be done in typed templates.
So we accept a symbol that was generated by a template/generic and has
an `{.inject.}` pragma and add it to scope again. For delayed syms this
is done by just setting the status back to `OkNew` from
`OkExistingFresh` so that `addSym` can define it, but another status
kind like `OkExistingInject` can be added for this. For procs the symbol
is manually added after pragmas, but could be done earlier with a
syntactic scan for `inject` as in untyped templates.

There was an issue with enum types defined in a scope: the `$` overload
for them is still generated in the top module scope but this means it
may not be able to load the enum symbol if it doesn't have a global sym
module qualifier, so type symbols now always generate a global sym.
(i.e. instead of generating `Foo.0` in a scope and `dollar.Foo.0` at top
level, it now uses `Foo.0.tabcdef`)

Enum fields are also never gensym'd in original Nim so they are
considered always injected here as well. (This caused redefinition
errors before overloadable enums.)

Something to note is that the symbols originally defined in the template
are still different from the substituted symbols in the instantiation,
even if they are fully typed. Injecting does not change this.