don't instantiate converters in generics + infix prefix etc stmt fix (#1119)
fixes #1117

Just like how normal calls are not instantiated in generics and produce
an explicit instantiation, do the same for generic converters (which
works because `HcallX` calls `semCall` as usual). In the test this is
encountered because the `items` call for `seq` produces a converter to
`openArray`.

Also the compiler has statement tags for `call` and `cmd` but not
`infix` and `prefix` etc, which is fixed by adding statement tags for
these and using a set `CallKindsS` instead of `CallKinds`.