implement `is`, make `when` work with unresolved condition (#1135)
closes #1125, refs #254, #252

The `is` operator is implemented, it semchecks to `true` or `false` if
both of the types are not generic, otherwise it stays as an `(is)` tree.
This means `is` with generic types cannot be evaluated at compile time.
So for `when T is int` etc to work, we also fix `when` so that it works
with unresolved conditions.

If a `when` condition cannot evaluate to `true` or `false` (also
includes the case where it produces an error), as long as we are in a
generic context, it does not produce an error saying that it cannot
evaluate the condition, instead it treats the whole `when` statement as
unresolved and leaves its branches untyped (semchecking them as untyped
as well via `semTemplBody`).

However as it is right now, this means if we are in generics and a
`when` condition cannot be resolved at compile time for other reasons,
it will not produce an error until the generic is instantiated. i.e. the
compiler does not distinguish between `when T is int` and `when
runtimeValue == 123` during typed generics. This might not be that big
of a deal but if we need to fix it in the long run we need to
distinguish between expressions that can't be evaluated due to
containing generics vs other reasons, related #252 (which points out
producing `untyped` type for the generic expressions would be too much).