fixes Hexer generates invalid Cursor (#1137)
In `trAsNamedType` proc in `hexer/nifcgen.nim`, a type declaration for
NifC is created and it is passed to `programs.publish` proc.
When `trType` proc is called with the symbol, that type declaration is
passed to `asTypeDecl` proc `in nimony/decls.nim`.
But that proc assumes given type declaration is `Nim-gear2`.
As type declaration for NifC has fewer child nodes, `takeTypeDecl` proc
produces invalid `Cursor`.