untyped args (#1136)
closes #1114

An argument that matches an `untyped` parameter passes the original tree
of the argument before semchecking rather than the semchecked version
with errors to the call. The argument is still semchecked, the result
and errors are just ignored. The original cursor is stored in an `orig:
Cursor` field inside a new `CallArg` type which is now used instead of
`Item` for call args.

The shadow scope rules were apparently simpler than I thought, original
Nim just never merges the shadow scope when matching a macro/template
(which makes sense, don't know what else I was thinking). So we don't
need a special typed/untyped distinction here for that.

Seems to make `defined`/`declared` work better as well, looked if there
was special logic for them to remove but there does not seem to be much.
`defined(abc.def)` works now after breaking in #1120.