fix typebound ops with nosystem and no system module in nimcache (#1147)
The string type becomes `string.0.sysvq0asl` even with `--noSystem`, so
typebound ops tries to load the system module to find its bound ops, but
this fails if the system module nif file does not exist yet. So this
triggered a random failure in #1145 when the nosystem category ran after
nimcache was wiped. Edit: And in master now in 2dea066.

To fix this, typebound ops just ignore types from the system module,
their bound ops should be in scope anyway. Nimcache is now also always
deleted before the nosystem category to test that this works.