propagate used typevar count in nested invoke (#1171)
fixes #1167

Invocations track if semchecking their arguments adds to a counter of
used typevar symbols, to decide if the invocation is concrete or not so
that it can be instantiated. However it does this by swapping the
counter with a local variable, which isolates the typevar count in the
arguments from the typevar count of the invocation itself. So
invocations with generic parameters do not count as generic when nested
inside another invocation, which wrongly instantiates it. To fix this,
the typevar counter is not isolated, only the counter changing is
tracked.

Worst case `containsGenericParams` can be used on each of the arguments.