fixes #1209 (#1212)
`semProcBody` proc in `nimony/semdecls.nim` calls `semExpr` twice for
the last expression of the proc body.
If the last expression was `[]` or `{}` and `semExpr` was called without
`AllowEmpty`, `semBracket` or `semCurly` calls `buildErr` proc.
If nimsem was built with `-d:debug`, `buildErr` prints given error and
quit.
If it was build without `-d:debug`, `buildErr` outputs `ErrT` node and
doesnt quit. Then `semProcBody` removes that `ErrT` node and no compiler
error is reported.
That is why the compile error was generated only when `-d:debug` is
defined.

This PR calls `semExpr` with `AllowEmpty` in `semProcBody` so that
`semBracket` or `semCurly` doesn't calls `buildErr` when they got `[]`
or `{}`.