fixes #1395; forbids non-string-literals  conversions to cstring (#1396)
fixes #1395

I use `cast[string]` for encodings changes because it uses unsafe
pointer arithmetic while the original string is changed in a loop. It
probably make `toCstring` unfit for this case since address is needed
and the original string is changed (which may mean the `\0` is
eliminated). Since `iconv` API also takes a length, it might be just
fine