fixes dynlib pragmas when used from other modules (#1397)
This PR implements `dynlib pragmas` cross modules by inline procs. It is
not as efficient as it is by `extern` the var symbol. Since `initDynlib`
is generated for each module and only when dynlib calls are called in
this module, I'm not sure if I can `extern` the var symbol in `nifcgen`
since no `initDynlib` is initialized in the original module