fixes ambiguous tuple field names causing nifc type mixing (#1410)
```nim
  of DotC:
    var a = n.firstSon
    skip a # skip the object
    let fld = a
    result = getTypeImpl(m, fld)
```

In `nifc/typenav`, `getType` uses the field of objects/tuples to get the
type of the field. However, the tuple field name is not a unique
identifier, which makes `getType` choose an arbitrary type for the tuple
field, supposing there are multiple tuples existing. i.e. before this
PR, the pattern of the field name of tuples is the same (e.g. `fld.0`,
`fld.1`).

This PR generates the tuple field name by combining its type, which
makes the field name unique to its type so that `getType` can always
find the right type.

Found it when debugging `try: except:`