infer openarray type for empty array and seq literals (#1572)
fixes #1565

The openarray converters for arrays and seqs are special cased in the
compiler. When they match empty array/seq literals, the converter's
generic params are first inferred from the type of the parameter in the
routine being called (presumably `openArray[T]`), by matching the return
type of the converter to the parameter type.

Then, the argument is matched as normal to the converter, however the
produced argument from the match is going to populate the empty literal
with the generic parameter of the converter, which also has to be
instantiated *relative to the converter* (i.e. only substituting the
converter's generic parameter) before it can be instantiated as normal
in the call to the original routine.

A matching converter call is then given the argument type `auto` to pass
to sigmatch, so that sigmatch can check for these generated converter
calls and always match them. This is needed if the matched `openArray`
type is generic, since when the converter call is being built, none of
the generic parameters of the routine are known, and so the type stays
as `openArray[T]` which sigmatch wouldn't accept. It is known that the
generated converter call always matches, since it depends on the
parameter type: If `T` isn't set before, then it will not be set here,
and the final match will fail due to uninstantiated parameters; and if
it is set before, then the previously set value has to be used, i.e.
there is no chance of `T` binding to a conflicting type.

Also empty array/set literals that receive non-array/set types now try
to convert to the given type by giving the literal an `auto` type rather
than immediately erroring.