don't resem untyped ast for value return result assignment (#1599)
fixes #1597, refs #1076

Currently reverts #1076 and breaks the original example in the test, a
workaround is used instead (`result = (a: a)` instead of just the
expression `(a: a)`). See [PR comment
below](https://github.com/nim-lang/nimony/pull/1599#issuecomment-3621610417)
for ideas on how to fix that example, can attempt those instead.